refactor: split manager leads router into read and write modules
